# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "SessionController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Create an opaque browser session",
        "tags": ["auth"]
      },
      "get": {
        "operationId": "SessionController_check",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Check the current browser session",
        "tags": ["auth"]
      },
      "delete": {
        "operationId": "SessionController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke the current browser session",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Seshat Finance API",
    "description": "API para registrar e organizar informações financeiras.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}